+++ modules/json/src/cpp/JsonEncode.cpp
@@ -24,25 +24,33 @@
 // LICENCE_BLOCK_END
 //=============================================================================
 #define _CRT_SECURE_NO_WARNINGS
-#include "JsonEncode.hpp"
-#include "characters_encoding.hpp"
+#include <stdio.h>
 #include <boost/algorithm/string.hpp>
 #include <iomanip>
 #include <sstream>
+#include <cstring>
 #include "nlsConfig.h"
+#include "JsonEncode.hpp"
+#include "characters_encoding.hpp"
+//=============================================================================
+#ifdef _MSC_VER
+#define snwprintf _snwprintf
+#endif
 //=============================================================================
 namespace Nelson {
 //=============================================================================
-static std::string jsonString;
+#define BUFFER_SIZE 4096
+//=============================================================================
+static std::wstring jsonString;
 //=============================================================================
 static void
-json_append_char(char c)
+json_append_char(wchar_t c)
 {
     jsonString.push_back(c);
 }
 //=============================================================================
 static void
-json_append_string(const std::string& str)
+json_append_string(const std::wstring& str)
 {
     jsonString.append(str);
 }
@@ -86,44 +94,52 @@ isSupportedType(ArrayOf ValueToEncode)
     return false;
 }
 //=============================================================================
+template <class T>
+static std::wstring
+wstringFormat(const std::wstring& format, T value)
+{
+    wchar_t buff[BUFFER_SIZE];
+#if _MSC_VER
+    int r = snwprintf(buff, BUFFER_SIZE, format.c_str(), value);
+#else
+    int r = swprintf(buff, format.c_str(), value);
+#endif
+    buff[r < BUFFER_SIZE ? r : BUFFER_SIZE - 1] = 0;
+    return std::wstring(buff);
+}
+//=============================================================================
 static void
 encode_character(wchar_t ch)
 {
     switch (ch) {
-    case L'"':
+    case L'"': {
+        json_append_string(L"\\\"");
+    } break;
     case L'\\': {
-        json_append_char('\\');
-        std::wstring wstr;
-        wstr.push_back(ch);
-        std::string str = wstring_to_utf8(wstr);
-        json_append_string(str);
+        json_append_string(L"\\\\");
     } break;
     case L'\b':
-        json_append_string("\\b");
+        json_append_string(L"\\b");
         break;
     case L'\f':
-        json_append_string("\\f");
+        json_append_string(L"\\f");
         break;
     case L'\n':
-        json_append_string("\\n");
+        json_append_string(L"\\n");
         break;
     case L'\r':
-        json_append_string("\\r");
+        json_append_string(L"\\r");
         break;
     case L'\t':
-        json_append_string("\\t");
+        json_append_string(L"\\t");
         break;
     default: {
         if ((ch > 13) && (ch < 32)) {
-            char buff[1024];
-            snprintf(buff, sizeof(buff), "\\u%04hx", ch);
-            std::string buffAsStdStr = buff;
-            json_append_string(buffAsStdStr);
+            json_append_string(wstringFormat<wchar_t>(L"\\u%04hx", ch));
         } else {
             std::wstring wstr;
             wstr.push_back(ch);
-            std::string str = wstring_to_utf8(wstr);
-            json_append_string(str);
+            json_append_string(wstr);
         }
     } break;
     }
@@ -134,25 +150,25 @@ encode_double(double val, bool convertNanInf)
 {
     if (std::isnan(val)) {
         if (convertNanInf) {
-            json_append_string("NaN,");
+            json_append_string(L"NaN,");
         } else {
-            json_append_string("null,");
+            json_append_string(L"null,");
         }
     } else if (std::isinf(val)) {
         if (convertNanInf) {
             if (val > 0) {
-                json_append_string("Inf,");
+                json_append_string(L"Inf,");
             } else {
-                json_append_string("-Inf,");
+                json_append_string(L"-Inf,");
             }
         } else {
-            json_append_string("null,");
+            json_append_string(L"null,");
         }
     } else {
-        std::stringstream stream;
+        std::wstringstream stream;
         stream << std::setprecision(std::numeric_limits<double>::digits10 + 1) << val;
-        std::string s = stream.str();
-        json_append_string(s + ",");
+        std::wstring s = stream.str();
+        json_append_string(s + L",");
     }
 }
 //=============================================================================
@@ -161,19 +177,19 @@ encode_single(single val, bool convertNanInf)
 {
     if (std::isnan(val)) {
         if (convertNanInf) {
-            json_append_string("NaN,");
+            json_append_string(L"NaN,");
         } else {
-            json_append_string("null,");
+            json_append_string(L"null,");
         }
     } else if (std::isinf(val)) {
         if (convertNanInf) {
             if (val > 0) {
-                json_append_string("Inf,");
+                json_append_string(L"Inf,");
             } else {
-                json_append_string("-Inf,");
+                json_append_string(L"-Inf,");
             }
         } else {
-            json_append_string("null,");
+            json_append_string(L"null,");
         }
     } else {
         encode_double(static_cast<double>(val), convertNanInf);
@@ -185,25 +201,25 @@ encode_array(ArrayOf ValueToEncode, bool close)
 {
     if (ValueToEncode.isCell()) {
         if (close) {
-            json_append_char(']');
+            json_append_char(L']');
         } else {
-            json_append_char('[');
+            json_append_char(L'[');
         }
     } else {
         indexType nbElements = ValueToEncode.getDimensions().getElementCount();
         if (nbElements > 1) {
             if (ValueToEncode.getDataClass() != NLS_CHAR) {
                 if (close) {
-                    json_append_char(']');
+                    json_append_char(L']');
                 } else {
-                    json_append_char('[');
+                    json_append_char(L'[');
                 }
             } else {
                 if (!ValueToEncode.isRowVector() && !ValueToEncode.isColumnVector()) {
                     if (close) {
-                        json_append_char(']');
+                        json_append_char(L']');
                     } else {
-                        json_append_char('[');
+                        json_append_char(L'[');
                     }
                 }
             }
@@ -213,16 +229,13 @@ encode_array(ArrayOf ValueToEncode, bool close)
 //=============================================================================
 template <class T>
 static void
-jsonEncodeInteger(const ArrayOf& ValueToEncode, std::string format)
+jsonEncodeInteger(const ArrayOf& ValueToEncode, const std::wstring& format)
 {
     auto* ptr = (T*)ValueToEncode.getDataPointer();
     if (ValueToEncode.isRowVector() || ValueToEncode.isColumnVector()) {
         indexType elementCount = ValueToEncode.getDimensions().getElementCount();
         for (indexType i = 0; i < elementCount; i++) {
-            char buff[1024];
-            snprintf(buff, sizeof(buff), format.c_str(), ptr[i]);
-            std::string buffAsStdStr = buff;
-            json_append_string(buffAsStdStr);
+            json_append_string(wstringFormat<T>(format, ptr[i]));
         }
     } else if (ValueToEncode.is2D()) {
         indexType rows = ValueToEncode.getDimensions().getRows();
@@ -230,15 +243,12 @@ jsonEncodeInteger(const ArrayOf& ValueToEncode, std::string format)
         for (int i = 0; i < rows; ++i) {
             json_append_char('[');
             for (int j = 0; j < cols; ++j) {
-                char buff[1024];
-                snprintf(buff, sizeof(buff), format.c_str(), ptr[j * rows + i]);
-                std::string buffAsStdStr = buff;
-                json_append_string(buffAsStdStr);
+                json_append_string(wstringFormat<T>(format, ptr[j * rows + i]));
             }
             if (boost::algorithm::ends_with(jsonString, L",")) {
                 jsonString.pop_back();
             }
-            json_append_string("],");
+            json_append_string(L"],");
         }
     } else {
         Dimensions dims = ValueToEncode.getDimensions();
@@ -247,15 +257,12 @@ jsonEncodeInteger(const ArrayOf& ValueToEncode, std::string format)
         for (int i = 0; i < ymax; ++i) {
             json_append_char('[');
             for (int j = 0; j < lastdimlen; ++j) {
-                char buff[1024];
-                snprintf(buff, sizeof(buff), format.c_str(), ptr[j * ymax + i]);
-                std::string buffAsStdStr = buff;
-                json_append_string(buffAsStdStr);
+                json_append_string(wstringFormat<T>(format, ptr[j * ymax + i]));
             }
             if (boost::algorithm::ends_with(jsonString, L",")) {
                 jsonString.pop_back();
             }
-            json_append_string("],");
+            json_append_string(L"],");
         }
     }
 }
@@ -273,28 +280,28 @@ jsonEncodeDouble(const ArrayOf& ValueToEncode, bool convertNanInf)
         indexType rows = ValueToEncode.getDimensions().getRows();
         indexType cols = ValueToEncode.getDimensions().getColumns();
         for (int i = 0; i < rows; ++i) {
-            json_append_char('[');
+            json_append_char(L'[');
             for (int j = 0; j < cols; ++j) {
                 encode_double(ptr[j * rows + i], convertNanInf);
             }
             if (boost::algorithm::ends_with(jsonString, L",")) {
                 jsonString.pop_back();
             }
-            json_append_string("],");
+            json_append_string(L"],");
         }
     } else {
         Dimensions dims = ValueToEncode.getDimensions();
         indexType lastdimlen = dims.getDimensionLength(dims.getLength() - 1);
         indexType ymax = dims.getElementCount() / lastdimlen;
         for (int i = 0; i < ymax; ++i) {
-            json_append_char('[');
+            json_append_char(L'[');
             for (int j = 0; j < lastdimlen; ++j) {
                 encode_double(ptr[j * ymax + i], convertNanInf);
             }
             if (boost::algorithm::ends_with(jsonString, L",")) {
                 jsonString.pop_back();
             }
-            json_append_string("],");
+            json_append_string(L"],");
         }
     }
 }
@@ -319,21 +326,21 @@ jsonEncodeSingle(const ArrayOf& ValueToEncode, bool convertNanInf)
             if (boost::algorithm::ends_with(jsonString, L",")) {
                 jsonString.pop_back();
             }
-            json_append_string("],");
+            json_append_string(L"],");
         }
     } else {
         Dimensions dims = ValueToEncode.getDimensions();
         indexType lastdimlen = dims.getDimensionLength(dims.getLength() - 1);
         indexType ymax = dims.getElementCount() / lastdimlen;
         for (int i = 0; i < ymax; ++i) {
-            json_append_char('[');
+            json_append_char(L'[');
             for (int j = 0; j < lastdimlen; ++j) {
                 encode_single(ptr[j * ymax + i], convertNanInf);
             }
             if (boost::algorithm::ends_with(jsonString, L",")) {
                 jsonString.pop_back();
             }
-            json_append_string("],");
+            json_append_string(L"],");
         }
     }
 }
@@ -344,31 +351,31 @@ jsonEncodeCharacters(const ArrayOf& ValueToEncode)
     std::wstring strw = ValueToEncode.getContentAsArrayOfCharacters();
     if (ValueToEncode.isRowVector() || ValueToEncode.isColumnVector()) {
         jsonString.reserve(jsonString.size() + strw.size() + 2);
-        json_append_char('"');
+        json_append_char(L'"');
         for (wchar_t i : strw) {
             encode_character(i);
         }
-        json_append_char('"');
+        json_append_char(L'"');
     } else if (ValueToEncode.is2D()) {
         indexType rows = ValueToEncode.getDimensions().getRows();
         indexType cols = ValueToEncode.getDimensions().getColumns();
         jsonString.reserve(jsonString.size() + strw.size() + 2);
         for (int i = 0; i < rows; ++i) {
-            json_append_char('"');
+            json_append_char(L'"');
             for (int j = 0; j < cols; ++j) {
                 encode_character(strw[j * rows + i]);
             }
             if (boost::algorithm::ends_with(jsonString, L",")) {
                 jsonString.pop_back();
             }
-            json_append_string("\",");
+            json_append_string(L"\",");
         }
     } else {
         Dimensions dims = ValueToEncode.getDimensions();
         indexType lastdimlen = dims.getDimensionLength(dims.getLength() - 1);
         indexType ymax = dims.getElementCount() / lastdimlen;
         for (int i = 0; i < ymax; ++i) {
-            json_append_char('\"');
+            json_append_char(L'\"');
             for (int j = 0; j < lastdimlen; ++j) {
                 wchar_t ch = strw[i * ymax + j];
                 encode_character(ch);
@@ -376,7 +383,7 @@ jsonEncodeCharacters(const ArrayOf& ValueToEncode)
             if (boost::algorithm::ends_with(jsonString, L",")) {
                 jsonString.pop_back();
             }
-            json_append_string("\",");
+            json_append_string(L"\",");
         }
     }
 }
@@ -389,9 +396,9 @@ jsonEncodeLogical(const ArrayOf& ValueToEncode)
         ompIndexType elementCount = ValueToEncode.getDimensions().getElementCount();
         for (ompIndexType i = 0; i < elementCount; i++) {
             if (ptr[i] == 0) {
-                json_append_string("false,");
+                json_append_string(L"false,");
             } else {
-                json_append_string("true,");
+                json_append_string(L"true,");
             }
         }
     } else if (ValueToEncode.is2D()) {
@@ -401,15 +408,15 @@ jsonEncodeLogical(const ArrayOf& ValueToEncode)
             json_append_char('[');
             for (indexType j = 0; j < cols; ++j) {
                 if (ptr[i] == 0) {
-                    json_append_string("false,");
+                    json_append_string(L"false,");
                 } else {
-                    json_append_string("true,");
+                    json_append_string(L"true,");
                 }
             }
             if (boost::algorithm::ends_with(jsonString, L",")) {
                 jsonString.pop_back();
             }
-            json_append_string("],");
+            json_append_string(L"],");
         }
     } else {
         Dimensions dims = ValueToEncode.getDimensions();
@@ -419,15 +426,15 @@ jsonEncodeLogical(const ArrayOf& ValueToEncode)
             json_append_char('[');
             for (indexType j = 0; j < lastdimlen; ++j) {
                 if (ptr[i] == 0) {
-                    json_append_string("false,");
+                    json_append_string(L"false,");
                 } else {
-                    json_append_string("true,");
+                    json_append_string(L"true,");
                 }
             }
             if (boost::algorithm::ends_with(jsonString, L",")) {
                 jsonString.pop_back();
             }
-            json_append_string("],");
+            json_append_string(L"],");
         }
     }
 }
@@ -441,9 +448,9 @@ jsonEncodeInternal(ArrayOf ValueToEncode, bool convertNanInf, std::wstring& erro
     }
     if (ValueToEncode.isEmpty()) {
         if (ValueToEncode.getDataClass() == NLS_CHAR) {
-            json_append_string("\"\"");
+            json_append_string(L"\"\"");
         } else {
-            json_append_string("[]");
+            json_append_string(L"[]");
         }
     } else {
         encode_array(ValueToEncode, false);
@@ -461,55 +468,55 @@ jsonEncodeInternal(ArrayOf ValueToEncode, bool convertNanInf, std::wstring& erro
             indexType elementCount = ValueToEncode.getDimensions().getElementCount();
             for (indexType i = 0; i < elementCount; i++) {
                 jsonEncodeInternal(elements[i], convertNanInf, errorMessage);
-                json_append_char(',');
+                json_append_char(L',');
             }
         } break;
         case NLS_STRUCT_ARRAY: {
             stringVector fieldnames = ValueToEncode.getFieldNames();
             indexType elementCount = ValueToEncode.getDimensions().getElementCount();
             for (int i = 0; i < elementCount; i++) {
-                json_append_char('{');
+                json_append_char(L'{');
                 for (auto fieldname : fieldnames) {
                     ArrayOfVector values = ValueToEncode.getFieldAsList(fieldname);
                     if (!values.empty()) {
-                        json_append_string("\"" + fieldname + "\":");
+                        json_append_string(L"\"" + utf8_to_wstring(fieldname) + L"\":");
                         jsonEncodeInternal(values[i], convertNanInf, errorMessage);
-                        json_append_char(',');
+                        json_append_char(L',');
                     }
                 }
                 if (boost::algorithm::ends_with(jsonString, L",")) {
                     jsonString.pop_back();
                 }
-                json_append_char('}');
-                json_append_char(',');
+                json_append_char(L'}');
+                json_append_char(L',');
             }
         } break;
         case NLS_LOGICAL: {
             jsonEncodeLogical(ValueToEncode);
         } break;
         case NLS_UINT8: {
-            jsonEncodeInteger<uint8>(ValueToEncode, "%u,");
+            jsonEncodeInteger<uint8>(ValueToEncode, L"%u,");
         } break;
         case NLS_INT8: {
-            jsonEncodeInteger<int8>(ValueToEncode, "%i,");
+            jsonEncodeInteger<int8>(ValueToEncode, L"%i,");
         } break;
         case NLS_UINT16: {
-            jsonEncodeInteger<uint16>(ValueToEncode, "%u,");
+            jsonEncodeInteger<uint16>(ValueToEncode, L"%u,");
         } break;
         case NLS_INT16: {
-            jsonEncodeInteger<int16>(ValueToEncode, "%i,");
+            jsonEncodeInteger<int16>(ValueToEncode, L"%i,");
         } break;
         case NLS_UINT32: {
-            jsonEncodeInteger<uint32>(ValueToEncode, "%u,");
+            jsonEncodeInteger<uint32>(ValueToEncode, L"%u,");
         } break;
         case NLS_INT32: {
-            jsonEncodeInteger<int32>(ValueToEncode, "%i,");
+            jsonEncodeInteger<int32>(ValueToEncode, L"%i,");
         } break;
         case NLS_UINT64: {
-            jsonEncodeInteger<uint64>(ValueToEncode, "%llu,");
+            jsonEncodeInteger<uint64>(ValueToEncode, L"%llu,");
         } break;
         case NLS_INT64: {
-            jsonEncodeInteger<uint64>(ValueToEncode, "%lli,");
+            jsonEncodeInteger<uint64>(ValueToEncode, L"%lli,");
         } break;
         case NLS_SINGLE: {
             jsonEncodeSingle(ValueToEncode, convertNanInf);